Support for negative fetish attributes
Completely overhauled dattAttributesAPIQuest
and
dattAttributeTrackerQuest.
Most of the code was outsourced into a
new script called
dattAttributesBaseQuest. API now mostly consist only
of functions with
minimal code and instead calls functions in BaseQuest.
APIQuest is
intended to be used by mods that rely on this mod as a hard
dependency.
TrackerQuest on the other hand only has events that call
the
corresponding functions in APIQuest. It is mainly used for mods that
do
not rely on this mod as a hard depency.
BaseQuest holds functions for
"working" with attributes. That means it
has functions for
changing/modifying those, or many other helpfull stuff
like validation
checks and convertions.

I've also rearranged dattConfigQuest again. It
should be now easier to
read and maintain. I've also added a new
function for the debug page to
fire events for stat changes mentioned
above.

Completely changed the naming pattern for variables in APIQuest
and
Tracker Quest, as well as changed some in ConfigQuest. If you do
not
like this change, we can go back to the old pattern thought.